ml/xray_classifier/requirements.txt
# X-ray Classifier Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0
pydantic>=2.0.0

# Optional: For production deployment
# fastapi>=0.100.0
# uvicorn>=0.20.0

# Optional: For enhanced functionality
# opencv-python>=4.5.0
# scikit-learn>=1.0.0
# matplotlib>=3.5.0

# For data handling
# pandas>=1.3.0

# For model training (if needed)
# tensorboard>=2.8.0
# wandb>=0.13.0
